+++ cocotb/result.py
@@ -31,23 +31,6 @@ from io import StringIO
 """Exceptions and functions for simulation result handling."""
 
 
-class ReturnValue(Exception):
-    """
-    Helper exception needed for Python versions prior to 3.3.
-
-    .. deprecated:: 1.4
-        Use a :keyword:`return` statement instead; this works in all supported versions of Python.
-    """
-
-    def __init__(self, retval):
-        warnings.warn(
-            "``ReturnValue`` is deprecated, use a normal return statement instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        self.retval = retval
-
-
 class TestComplete(Exception):
     """
     Exception showing that the test was completed. Sub-exceptions detail the exit status.